refactor:  update component styles
