+++ docs/package.json
@@ -30,7 +30,6 @@
     "@next/bundle-analyzer": "15.2.0",
     "canvas-confetti": "1.9.3",
     "clsx": "2.1.1",
-    "color2k": "2.0.3",
     "hast-util-to-jsx-runtime": "2.3.6",
     "next": "15.3.1",
     "nextra": "4.2.3",
