--- packages/react/src/theme/shared/colorVariant.ts
@@ -1,29 +1,0 @@
-export const colorVariant = {
-  blue: {
-    colorPalette: 'blue',
-  },
-  gray: {
-    colorPalette: 'gray',
-  },
-  pink: {
-    colorPalette: 'pink',
-  },
-  purple: {
-    colorPalette: 'purple',
-  },
-  cyan: {
-    colorPalette: 'cyan',
-  },
-  orange: {
-    colorPalette: 'orange',
-  },
-  yellow: {
-    colorPalette: 'yellow',
-  },
-  green: {
-    colorPalette: 'green',
-  },
-  red: {
-    colorPalette: 'red',
-  },
-} as const
