--- packages/react/src/theme/shared/fullWidthVariant.ts
@@ -1,5 +1,0 @@
-export const fullWidth = {
-  true: {
-    w: '100%',
-  },
-} as const
