--- packages/react/src/theme/shared/radiusVariant.ts
@@ -1,17 +1,0 @@
-export const radiusVariant = {
-  sm: {
-    borderRadius: 'md',
-  },
-  md: {
-    borderRadius: 'lg',
-  },
-  lg: {
-    borderRadius: 'xl',
-  },
-  full: {
-    borderRadius: 'full',
-  },
-  none: {
-    borderRadius: 'none',
-  },
-} as const
