--- packages/react/src/theme/shared/sizeVariant.ts
@@ -1,17 +1,0 @@
-export const sizeVariant = {
-  sm: {
-    px: '3',
-    fs: 'md',
-    h: '8',
-  },
-  md: {
-    px: '4',
-    fs: 'lg',
-    h: '10',
-  },
-  lg: {
-    px: '5',
-    fs: 'xl',
-    h: '12',
-  },
-} as const
