+++ packages/react/src/theme/system/selectors.ts
@@ -2,6 +2,8 @@ import { defineConfig } from '@nex-ui/system'
 
 export const selectors = defineConfig.selectors({
   hover: '&:is(:hover, [data-hover=true]):not(:disabled, [data-disabled=true])',
+  checked:
+    '&:is(:checked, [data-checked=true]):not(:disabled, [data-disabled=true])',
   active:
     '&:is(:active, [data-active=true]):not(:disabled, [data-disabled=true])',
   focus: '&:is(:focus, [data-focus=true]):not(:disabled, [data-disabled=true])',
